requirements.txt
flask-migrate~=4.0.7
SQLAlchemy~=2.0.30
marvel~=0.1.0
python-dotenv~=1.0.1
WTForms~=3.1.2
PyYAML~=6.0.1
Flask~=3.0.3
alembic~=1.13.1
flask-sqlalchemy~=3.1.1
flask-bcrypt~=1.0.1
psycopg2-binary~=2.9.9
flask-wtf~=1.2.1
flask_login~=0.6.3